Fixed the issue with numbers being repositioned due to non-stable sorts.
Since numbers don't affect alphabet sorting, they can just be filtered
out of the array, then the alphabets are sorted and the numbers tacked
back on at the end.
This solves the positional issue and removes the need for one of the
sorting functions (the one that deals with mixed numbers and alphabets)
entirely.
All test cases pass now.